Number Guessing Game Using JavaScript
This is Number guessing game and I have used JavaScript for calculation and I used if-else function for provide the winner after inputing right number and The number is fixed 1 to 10. Now I have used switch case for showing the progress bar and maximum input trying is 3 . that's mean, user should currectly guess within 3 times otherwise it will be overed and say to start again the game.